@@ -1,41 +1,60 @@
 {
-  "name": "raiym/instagram-php-scraper",
-  "description": "Instagram PHP Scraper. Get account information, photos and videos without any authorization",
-  "keywords": [
-    "instagram",
-    "scraper"
-  ],
-  "type": "library",
-  "homepage": "https://github.com/raiym/instagram-php-scraper",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "raiym",
-      "email": "r.kapishev@gmail.com"
+    "name": "raiym/instagram-php-scraper",
+    "description": "Instagram PHP Scraper. Get account information, photos and videos without any authorization",
+    "keywords": [
+        "instagram",
+        "scraper"
+    ],
+    "type": "library",
+    "homepage": "https://github.com/raiym/instagram-php-scraper",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "raiym",
+            "email": "r.kapishev@gmail.com"
+        }
+    ],
+    "require": {
+        "php": "^7.1.3",
+        "mashape/unirest-php": "3.0.*",
+        "ext-curl": "*",
+        "ext-json": "*",
+        "psr/simple-cache": "~1.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^7.5",
+        "phpfastcache/phpfastcache": "^7.1",
+        "rector/rector": "^0.6.13",
+        "friendsofphp/php-cs-fixer": "^2.16"
+    },
+    "config": {
+        "preferred-install": {
+            "*": "dist"
+        },
+        "sort-packages": true
+    },
+    "autoload": {
+        "psr-4": {
+            "InstagramScraper\\": "src/InstagramScraper/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "InstagramScraper\\Tests\\": "tests/"
+        }
+    },
+    "support": {
+        "email": "r.kapishev@gmail.com"
+    },
+    "scripts": {
+        "beautify": [
+            "./vendor/bin/php-cs-fixer fix --verbose --config=.php_cs.dist src/"
+        ],
+        "beautify-dry-run": [
+            "./vendor/bin/php-cs-fixer fix --diff --dry-run -v --config=.php_cs.dist src/"
+        ],
+        "rector": [
+            "./vendor/bin/rector process src"
+        ]
     }
-  ],
-  "require": {
-    "php": ">=5.4.0",
-    "mashape/unirest-php": "3.0.*",
-    "ext-curl": "*",
-    "ext-json": "*",
-    "psr/simple-cache": "~1.0"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "5.5.*",
-    "phpfastcache/phpfastcache": "^7.1"
-  },
-  "autoload": {
-    "psr-4": {
-      "InstagramScraper\\": "src/InstagramScraper/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "InstagramScraper\\Tests\\": "tests/"
-    }
-  },
-  "support": {
-    "email": "r.kapishev@gmail.com"
-  }
 }
