@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace InstagramScraper;
 
@@ -8,8 +9,8 @@ class InstagramQueryId
     /**
      * id = {{accoundId}}, first = {{count}}, after = {{mediaId}}
      */
-    const USER_MEDIAS = '17880160963012870';
-    const USER_STORIES = '17890626976041463';
-    const STORIES = '17873473675158481';
+    public const USER_MEDIAS = '17880160963012870';
+    public const USER_STORIES = '17890626976041463';
+    public const STORIES = '17873473675158481';
 
 }
